Provide adapter for use with `jk generate`
Now that templates can return objects, rather than string chunks, the
value of write.js is the organisation of objects into files -- which
is pretty close to what `jk generate` requires. So, co-opt write.js as
generate.js, and use it to provide a version of `chart` that can be
used directly with `jk generate`.
